Move around persistence code
