+++ pigdef/Cargo.toml
@@ -22,8 +22,6 @@ iroh = []
 
 [dependencies]
 serde = { version = "1.0.217", default-features = false, features = ["derive"] }
-# for config serialization to/from file
-serde_json = { version = "1.0.138", default-features = false, features = ["std"] }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
 tempfile = "3"
